Extract compact help overlay builder
