+++ src/Func/Commands/Start/Dashboard/Rendering/CompactRenderer.cs
@@ -37,6 +37,7 @@ internal sealed class CompactRenderer(
     private readonly IAnsiConsole _console = console ?? AnsiConsole.Console;
     private readonly CompactHeaderBuilder _headerBuilder = new(interaction.Theme, runInfo ?? new());
     private readonly CompactFooterBuilder _footerBuilder = new(interaction.Theme, runInfo ?? new());
+    private readonly CompactHelpOverlayBuilder _helpOverlayBuilder = new(interaction.Theme);
     private readonly Lock _uiLock = new();
     private readonly CompactLogBuffer _logBuffer = new(MaxLogTailLines);
     private readonly CompactLogLineFormatter _logLineFormatter = new(interaction.Theme, palette);
@@ -64,7 +65,6 @@ internal sealed class CompactRenderer(
     private ITheme Theme => _interaction.Theme;
 
     private string HeadingTag => field ??= Theme.Heading.ToMarkup();
-    private string CommandTag => field ??= Theme.Command.ToMarkup();
     private string MutedTag => field ??= Theme.Muted.ToMarkup();
     private string EmphasisTag => field ??= Theme.Emphasis.ToMarkup();
     private string SuccessTag => field ??= Theme.Success.ToMarkup();
@@ -824,42 +824,7 @@ internal sealed class CompactRenderer(
     }
 
     private IRenderable BuildHelpOverlay()
-    {
-        Table table = new Table()
-            .Border(TableBorder.None)
-            .HideHeaders()
-            .Expand()
-            .AddColumn(new TableColumn(string.Empty).PadLeft(1).PadRight(3).NoWrap())
-            .AddColumn(new TableColumn(string.Empty).PadRight(0));
-
-        AddHelpRow(table, "?", "Toggle this help panel.");
-        AddHelpRow(table, "t", "Open the function browser.");
-        AddHelpRow(table, "/", "Search functions by name, trigger, or route.");
-        AddHelpRow(table, "↑/↓", "Move selection in the function browser.");
-        AddHelpRow(table, "PgUp/PgDn", "Scroll logs; in the function browser, jump through functions.");
-        AddHelpRow(table, "Home/End", "Jump to oldest logs / latest logs, or first / last function in the browser.");
-        AddHelpRow(table, "Enter", "Filter logs to the selected function in the function browser.");
-        AddHelpRow(table, "a", "Clear the active function filter.");
-        AddHelpRow(table, "f", "Cycle the active function filter.");
-        AddHelpRow(table, "c", "Clear visible log output.");
-        AddHelpRow(table, "e", "Toggle errors-only log view.");
-        AddHelpRow(table, "1/2/3", "Set visible log level: info, warning, or error.");
-        AddHelpRow(table, "q", "Stop the host.");
-        AddHelpRow(table, "Esc", "Close the active overlay.");
-        AddHelpRow(table, "Ctrl+C", "Stop the host.");
-
-        var panel = new Panel(new Rows(
-            new Markup($"[{MutedTag}]Available compact-mode controls[/]"),
-            table))
-        {
-            Header = new PanelHeader("Help"),
-            Border = BoxBorder.Rounded,
-            BorderStyle = Theme.Muted,
-            Expand = true,
-        };
-
-        return panel;
-    }
+        => _helpOverlayBuilder.Build();
 
     private IRenderable BuildFunctionSearch(DashboardSnapshot snapshot)
     {
@@ -951,13 +916,6 @@ internal sealed class CompactRenderer(
         }
 
         return table;
-    }
-
-    private void AddHelpRow(Table table, string key, string description)
-    {
-        table.AddRow(
-            new Markup($"[{CommandTag}]{Markup.Escape(key)}[/]"),
-            new Markup($"[{MutedTag}]{Markup.Escape(description)}[/]"));
     }
 
     private IRenderable BuildFunctionBrowser(DashboardSnapshot snapshot)
